requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
idna==3.10
imagecodecs
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mkl-service==2.4.0
mkl_fft
mkl_random
moviepy==2.1.1
mpmath==1.3.0
networkx==3.3
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
psutil==6.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn
scipy
seaborn==0.13.2
six==1.16.0
sympy==1.13.3
threadpoolctl
tifffile
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.103
ultralytics-thop==2.0.8
urllib3==2.2.3
